+++ pymc3/plots/__init__.py
@@ -5,3 +5,4 @@ from .kdeplot import kdeplot
 from .posteriorplot import plot_posterior, plot_posterior_predictive_glm
 from .traceplot import traceplot
 from .energyplot import energyplot
+from .densityplot import densityplot