+++ pymc3/step_methods/hmc/integration.py
@@ -7,6 +7,10 @@ from scipy import linalg
 State = namedtuple("State", 'q, p, v, q_grad, energy')
 
 
+class IntegrationError(RuntimeError):
+    pass
+
+
 class CpuLeapfrogIntegrator(object):
     def __init__(self, potential, logp_dlogp_func):
         """Leapfrog integrator using CPU."""
@@ -46,6 +50,21 @@ class CpuLeapfrogIntegrator(object):
         -------
         None if `out` is provided, else a State namedtuple
         """
+        try:
+            return self._step(epsilon, state, out=None)
+        except linalg.LinAlgError as err:
+            msg = "LinAlgError during leapfrog step."
+            raise IntegrationError(msg)
+        except ValueError as err:
+            # Raised by many scipy.linalg functions
+            scipy_msg = "array must not contain infs or nans"
+            if len(err.args) > 0 and scipy_msg in err.args[0].lower():
+                msg = "Infs or nans in scipy.linalg during leapfrog step."
+                raise IntegrationError(msg)
+            else:
+                raise
+
+    def _step(self, epsilon, state, out=None):
         pot = self._potential
         axpy = linalg.blas.get_blas_funcs('axpy', dtype=self._dtype)