+++ pymc3/tests/test_text_backend.py
@@ -12,7 +12,7 @@ class TestTextSampling(object):
         with pm.Model():
             pm.Normal("mu", mu=0, sd=1, shape=2)
             db = text.Text(self.name)
-            pm.sample(20, tune=10, init=None, trace=db)
+            pm.sample(20, tune=10, init=None, trace=db, njobs=2)
 
     def teardown_method(self):
         bf.remove_file_or_directory(self.name)