+++ scripts/create_testenv.sh
@@ -45,10 +45,10 @@ if [ ${PYTHON_VERSION} == "2.7" ]; then
     pip install mock
 fi
 
-pip install pytest pytest-cov nose-parameterized pylint
+pip install -r requirements-dev.txt
 
 # Install untested, non-required code (linter fails without them)
-pip install ipython ipywidgets numdifftools
+pip install ipython ipywidgets
 
 # matplotlib is not required for the library, but is for tests
 pip install matplotlib