+++ RELEASE-NOTES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## PyMC 3.3. (Unreleased) 
+## PyMC 3.3 (January 9, 2018)
 
 ### New features
 
@@ -14,6 +14,8 @@
   algorithm (#2544)
 - Michael Osthege added support for population-samplers and implemented differential evolution metropolis (`DEMetropolis`).  For models with correlated dimensions that can not use gradient-based samplers, the `DEMetropolis` sampler can give higher effective sampling rates. (also see [PR#2735](https://github.com/pymc-devs/pymc3/pull/2735))
 - Forestplot supports multiple traces (#2736)
+- Add new plot, densityplot (#2741)
+- DIC and BPIC calculations have been deprecated
 
 ### Fixes
 
@@ -24,6 +26,12 @@
 - Add test for `model.logp_array` and `model.bijection` (#2724) 
 - Fixed `sample_ppc` and `sample_ppc_w` to iterate all chains(#2633, #2748)
 - Add Bayesian R2 score (for GLMs) `stats.r2_score` (#2696) and test (#2729).
+- SMC works with transformed variables (#2755)
+- Speedup OPVI (#2759)
+
+### Deprecations
+
+- Old (`minibatch-`)`advi` is removed (#2781)
 
 
 ## PyMC3 3.2 (October 10, 2017)